Use API to link wallet in WalletLinkingTest
