+++ integrationtests/src/main/kotlin/xyz/funkybit/integrationtests/utils/Utils.kt
@@ -1,9 +1,23 @@
 package xyz.funkybit.integrationtests.utils
 
+import kotlinx.datetime.Clock
+import kotlinx.datetime.Instant
+import org.bitcoinj.core.ECKey
+import org.web3j.crypto.Credentials
+import org.web3j.crypto.ECKeyPair
 import xyz.funkybit.apps.api.model.Balance
+import xyz.funkybit.apps.api.model.BitcoinLinkAddressProof
+import xyz.funkybit.apps.api.model.EvmLinkAddressProof
 import xyz.funkybit.apps.api.model.SymbolInfo
 import xyz.funkybit.apps.api.model.websocket.Limits
+import xyz.funkybit.core.evm.ECHelper
+import xyz.funkybit.core.evm.EIP712Helper
+import xyz.funkybit.core.model.BitcoinAddress
+import xyz.funkybit.core.model.BitcoinSignature
+import xyz.funkybit.core.model.EvmAddress
+import xyz.funkybit.core.model.EvmSignature
 import xyz.funkybit.core.model.Symbol
+import xyz.funkybit.core.model.db.ChainId
 import xyz.funkybit.core.utils.fromFundamentalUnits
 import xyz.funkybit.core.utils.toFundamentalUnits
 import java.math.BigDecimal
@@ -87,4 +101,44 @@ fun assertFee(expectedAmount: AssetAmount, amount: BigInteger, symbol: Symbol) {
 
 fun verifyApiReturnsSameLimits(apiClient: TestApiClient, wsMessage: Limits) {
     assertEquals(wsMessage.limits, apiClient.getLimits().limits)
+}
+
+fun signBitcoinWalletLinkProof(
+    ecKey: ECKey,
+    address: BitcoinAddress,
+    linkAddress: EvmAddress,
+    timestamp: Instant = Clock.System.now(),
+): BitcoinLinkAddressProof {
+    val message = "[funkybit] Please sign this message to link your wallets. This action will not cost any gas fees."
+    val bitcoinLinkAddressMessage = "$message\nAddress: ${address.value}, LinkAddress: ${linkAddress.value}, Timestamp: $timestamp"
+    val signature = BitcoinSignature(ecKey.signMessage(bitcoinLinkAddressMessage))
+
+    return BitcoinLinkAddressProof(
+        message = message,
+        address = address,
+        linkAddress = linkAddress,
+        timestamp = timestamp.toString(),
+        signature = signature,
+    )
+}
+
+fun signEvmWalletLinkProof(
+    ecKeyPair: ECKeyPair,
+    address: EvmAddress,
+    linkAddress: BitcoinAddress,
+    chainId: ChainId = ChainId(1337U),
+    timestamp: Instant = Clock.System.now(),
+): EvmLinkAddressProof {
+    val evmLinkAddressProof = EvmLinkAddressProof(
+        message = "[funkybit] Please sign this message to link your wallets. This action will not cost any gas fees.",
+        address = address,
+        linkAddress = linkAddress,
+        chainId = chainId,
+        timestamp = timestamp.toString(),
+        signature = EvmSignature.emptySignature(),
+    )
+
+    val signature: EvmSignature = ECHelper.signData(Credentials.create(ecKeyPair), EIP712Helper.computeHash(evmLinkAddressProof))
+
+    return evmLinkAddressProof.copy(signature = signature)
 }
