+++ integrationtests/src/test/kotlin/xyz/funkybit/integrationtests/api/IdentityApiTest.kt
@@ -1,26 +1,17 @@
 package xyz.funkybit.integrationtests.api
 
-import kotlinx.datetime.Clock
-import kotlinx.datetime.Instant
 import org.bitcoinj.core.ECKey
 import org.bitcoinj.core.NetworkParameters
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
-import org.web3j.crypto.Credentials
-import org.web3j.crypto.ECKeyPair
-import xyz.funkybit.apps.api.model.BitcoinLinkAddressProof
-import xyz.funkybit.apps.api.model.EvmLinkAddressProof
 import xyz.funkybit.apps.api.model.LinkIdentityApiRequest
-import xyz.funkybit.core.evm.ECHelper
-import xyz.funkybit.core.evm.EIP712Helper
 import xyz.funkybit.core.model.BitcoinAddress
-import xyz.funkybit.core.model.BitcoinSignature
 import xyz.funkybit.core.model.EvmAddress
-import xyz.funkybit.core.model.EvmSignature
-import xyz.funkybit.core.model.db.ChainId
 import xyz.funkybit.integrationtests.testutils.AppUnderTestRunner
 import xyz.funkybit.integrationtests.utils.TestApiClient
 import xyz.funkybit.integrationtests.utils.WalletKeyPair
+import xyz.funkybit.integrationtests.utils.signBitcoinWalletLinkProof
+import xyz.funkybit.integrationtests.utils.signEvmWalletLinkProof
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 
@@ -59,45 +50,5 @@ class IdentityApiTest {
     fun `error cases`() {
         // invalid signatures, timestamp, addresses
         // link address in already in use
-    }
-
-    private fun signBitcoinWalletLinkProof(
-        ecKey: ECKey,
-        address: BitcoinAddress,
-        linkAddress: EvmAddress,
-        timestamp: Instant = Clock.System.now(),
-    ): BitcoinLinkAddressProof {
-        val message = "[funkybit] Please sign this message to link your wallets. This action will not cost any gas fees."
-        val bitcoinLinkAddressMessage = "$message\nAddress: ${address.value}, LinkAddress: ${linkAddress.value}, Timestamp: $timestamp"
-        val signature = BitcoinSignature(ecKey.signMessage(bitcoinLinkAddressMessage))
-
-        return BitcoinLinkAddressProof(
-            message = message,
-            address = address,
-            linkAddress = linkAddress,
-            timestamp = timestamp.toString(),
-            signature = signature,
-        )
-    }
-
-    private fun signEvmWalletLinkProof(
-        ecKeyPair: ECKeyPair,
-        address: EvmAddress,
-        linkAddress: BitcoinAddress,
-        chainId: ChainId = ChainId(1337U),
-        timestamp: Instant = Clock.System.now(),
-    ): EvmLinkAddressProof {
-        val evmLinkAddressProof = EvmLinkAddressProof(
-            message = "[funkybit] Please sign this message to link your wallets. This action will not cost any gas fees.",
-            address = address,
-            linkAddress = linkAddress,
-            chainId = chainId,
-            timestamp = timestamp.toString(),
-            signature = EvmSignature.emptySignature(),
-        )
-
-        val signature: EvmSignature = ECHelper.signData(Credentials.create(ecKeyPair), EIP712Helper.computeHash(evmLinkAddressProof))
-
-        return evmLinkAddressProof.copy(signature = signature)
     }
 }
